Work-in-progress: Modularization of Simulator
