+++ setup.cfg
@@ -1,2 +1,3 @@
 [flake8]
 max-line-length = 120
+ignore = F401, F403, F405
