+++ src/coordsim/forwarders/link_load_forwarder.py
@@ -1,0 +1,9 @@
+class LinkLoadForwarder(object):
+    """
+    This is the default forwarder class. It takes the shortest path.
+    """
+    def __init__(self):
+        """
+        docstring
+        """
+        pass
